Fix encoding/decoding of small contents
The library refused to work with small contents with more than
one chunk of padding. Now the function encoding_prepare() still
returns 0 for these contents, but the returned structure is OK
for use with rain_encode() and rain_rehydrate().

The packet size has been optimized, and we have less cases of
"more than one chunk of padding".

The test_librain executable is now compiled by default.

Change-Id: I543cca55381713e45cbae478352ed48dd82cac6b